Iot_20180120 TeaDSL sdk.

+++ iot-20180120/core/Models/ListDistributedProductResponseBody.cs
@@ -27,7 +27,7 @@ namespace AlibabaCloud.SDK.Iot20180120.Models
                 public class ListDistributedProductResponseBodyDataInfoItems : TeaModel {
                     [NameInMap("GmtCreate")]
                     [Validation(Required=false)]
-                    public int? GmtCreate { get; set; }
+                    public long? GmtCreate { get; set; }
 
                     [NameInMap("ProductKey")]
                     [Validation(Required=false)]
